Done section 18

+++ Client/src/app/messages/messages.component.ts
@@ -1,3 +1,4 @@
+import { ConfirmService } from './../_services/confirm.service';
 import { MessageService } from './../_services/message.service';
 import { Pagination } from './../_models/pagination';
 import { Message } from './../_models/message';
@@ -16,7 +17,8 @@ export class MessagesComponent implements OnInit {
   pageSize = 5;
   loading = false;
 
-  constructor(private messageService: MessageService) { }
+  constructor(private messageService: MessageService,
+              private confirmService: ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages();
@@ -25,17 +27,23 @@ export class MessagesComponent implements OnInit {
   loadMessages() {
     this.loading = true;
     this.messageService
-    .getMessages(this.pageNumber, this.pageSize, this.container)
-    .subscribe(response  => {
-      this.messages = response.result;
-      this.pagination = response.pagination;
-      this.loading = false;
-    });
+      .getMessages(this.pageNumber, this.pageSize, this.container)
+      .subscribe(response => {
+        this.messages = response.result;
+        this.pagination = response.pagination;
+        this.loading = false;
+      });
   }
   deleteMessage(id: number) {
-    this.messageService.deleteMessage(id).subscribe(() => {
-      this.messages.splice(this.messages.findIndex( m => m.id === id ), 1);
-    });
+    this.confirmService.confirm('Confirm delete', 'This cannot be done')
+      .subscribe(result => {
+        if (result) {
+          this.messageService.deleteMessage(id).subscribe(() => {
+            this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
+          });
+        }
+      });
+
   }
   pageChanged(event: any) {
     this.pageNumber = event.page;
